python/requirements.txt
numpy>=1.24
PyQt6>=6.5
pyqtgraph>=0.13
psycopg2-binary>=2.9 
scipy>=1.10
torch>=2.0 